+++ sdmx-dl-api/src/test/java/sdmxdl/repo/SdmxRepositoryTest.java
@@ -32,7 +32,7 @@ public class SdmxRepositoryTest {
 
     @Test
     public void testBuilder() {
-        assertThat(SdmxRepository.builder().name("test").dataSet(dataSet).build().isDetailSupported()).isTrue();
+        assertThat(SdmxRepository.builder().name("test").dataSet(dataSet).build().getSupportedFeatures()).isEmpty();
     }
 
     @Test