+++ CHANGELOG.md
@@ -21,6 +21,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Refactor `SdmxManager` API
 - Replace `SdmxWebConnection#ping()` with `SdmxConnection#testConnection()`
 - Refactor data query API [#218](https://github.com/nbbrd/sdmx-dl/issues/218)
+- Replace `SdmxConnection#isDetailSupported()` with a more general solution [#89](https://github.com/nbbrd/sdmx-dl/issues/89)
 
 ### Fixed